requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
ecdsa==0.19.1
idna==3.10
pyasn1==0.6.1
PyJWT==2.10.1
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
